[#660] Split account logic classes into separate files
Migrated each account logic class to its dedicated file:
* Enhances code organization
* Improves maintainability and readability
* Simplifies future modifications and debugging
